Reworked server structure in Program.cs
